requirements.txt
Flask==2.3.2
PyYAML==6.0
requests==2.30.0
psutil==5.9.5
Flask-Cors==3.0.10